Various Refactors (#99)
* first round of refactors

* split Moderation.kt and Utilities.kt

* expand try blocks

* rename context-specific config functions for consistency
