+++ src/main/kotlin/net/irisshaders/lilybot/extensions/util/ModUtilities.kt
@@ -4,8 +4,6 @@ package net.irisshaders.lilybot.extensions.util
 
 import com.kotlindiscord.kord.extensions.DISCORD_BLACK
 import com.kotlindiscord.kord.extensions.DISCORD_BLURPLE
-import com.kotlindiscord.kord.extensions.DISCORD_GREEN
-import com.kotlindiscord.kord.extensions.DISCORD_YELLOW
 import com.kotlindiscord.kord.extensions.checks.hasPermission
 import com.kotlindiscord.kord.extensions.commands.Arguments
 import com.kotlindiscord.kord.extensions.commands.converters.impl.defaultingBoolean
@@ -13,7 +11,6 @@ import com.kotlindiscord.kord.extensions.commands.converters.impl.optionalChanne
 import com.kotlindiscord.kord.extensions.commands.converters.impl.string
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
-import com.kotlindiscord.kord.extensions.extensions.publicSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.entity.Permission
 import dev.kord.common.entity.PresenceStatus
@@ -21,59 +18,18 @@ import dev.kord.core.behavior.channel.GuildMessageChannelBehavior
 import dev.kord.core.behavior.channel.MessageChannelBehavior
 import dev.kord.core.behavior.channel.createEmbed
 import dev.kord.core.behavior.channel.createMessage
-import dev.kord.rest.builder.message.create.embed
 import dev.kord.rest.request.KtorRequestException
 import kotlinx.datetime.Clock
 import net.irisshaders.lilybot.utils.DatabaseHelper
-import net.irisshaders.lilybot.utils.ONLINE_STATUS_CHANNEL
 import net.irisshaders.lilybot.utils.TEST_GUILD_ID
-import net.irisshaders.lilybot.utils.getFromConfigPrivateResponse
+import net.irisshaders.lilybot.utils.getConfigPrivateResponse
 import net.irisshaders.lilybot.utils.responseEmbedInChannel
 import kotlin.time.ExperimentalTime
 
-class Utilities : Extension() {
-	override val name = "utilities"
+class ModUtilities : Extension() {
+	override val name = "mod-utilities"
 
 	override suspend fun setup() {
-		val onlineLog = kord.getGuild(TEST_GUILD_ID)?.getChannel(ONLINE_STATUS_CHANNEL) as GuildMessageChannelBehavior
-
-		/**
-		 * Online notification
-		 * @author IMS212
-		 */
-		responseEmbedInChannel(
-			onlineLog, "Lily is now online!", null,
-			DISCORD_GREEN, null
-		)
-
-		/**
-		 * Ping Command
-		 * @author IMS212
-		 * @author NoComment1105
-		 */
-		publicSlashCommand {  // Public slash commands have public responses
-			name = "ping"
-			description = "Am I alive?"
-
-			action {
-				val averagePing = this@Utilities.kord.gateway.averagePing
-
-				respond {
-					embed {
-						color = DISCORD_YELLOW
-						title = "Pong!"
-
-						timestamp = Clock.System.now()
-
-						field {
-							name = "Your Ping with Lily is:"
-							value = "**$averagePing**"
-							inline = true
-						}
-					}
-				}
-			}
-		}
 
 		/**
 		 * Say Command
@@ -86,7 +42,7 @@ class Utilities : Extension() {
 			check { hasPermission(Permission.ModerateMembers) }
 
 			action {
-				val actionLogId = getFromConfigPrivateResponse("modActionLog") ?: return@action
+				val actionLogId = getConfigPrivateResponse("modActionLog") ?: return@action
 
 				val actionLog = guild?.getChannel(actionLogId) as GuildMessageChannelBehavior
 				val targetChannel = if (arguments.targetChannel == null) {
@@ -142,7 +98,7 @@ class Utilities : Extension() {
 					return@action
 				}
 
-				val actionLogId = getFromConfigPrivateResponse("modActionLog") ?: return@action
+				val actionLogId = getConfigPrivateResponse("modActionLog") ?: return@action
 
 				val actionLog = guild?.getChannel(actionLogId) as GuildMessageChannelBehavior
 
@@ -151,7 +107,7 @@ class Utilities : Extension() {
 					playing(arguments.presenceArgument)
 				}
 
-				DatabaseHelper.putInStatus(arguments.presenceArgument)
+				DatabaseHelper.setStatus(arguments.presenceArgument)
 
 				respond { content = "Presence set to `${arguments.presenceArgument}`" }
 
@@ -169,15 +125,15 @@ class Utilities : Extension() {
 	inner class SayArgs : Arguments() {
 		val messageArgument by string {
 			name = "message"
-			description = "Message contents"
+			description = "The text of the message to be sent"
 		}
 		val targetChannel by optionalChannel {
 			name = "channel"
-			description = "The channel you want to send the message in"
+			description = "The channel the message should be sent in"
 		}
 		val embedMessage by defaultingBoolean {
 			name = "embed"
-			description = "Would you like to send as embed"
+			description = "If the message should be sent as an embed"
 			defaultValue = false
 		}
 	}
@@ -185,7 +141,7 @@ class Utilities : Extension() {
 	inner class PresenceArgs : Arguments() {
 		val presenceArgument by string {
 			name = "presence"
-			description = "Lily's presence"
+			description = "The new value Lily's presence should be set to"
 		}
 	}
 }
