+++ src/main/kotlin/net/irisshaders/lilybot/extensions/util/RoleMenu.kt
@@ -27,7 +27,7 @@ import dev.kord.rest.builder.message.create.embed
 import kotlinx.datetime.Clock
 import net.irisshaders.lilybot.utils.ComponentData
 import net.irisshaders.lilybot.utils.DatabaseHelper
-import net.irisshaders.lilybot.utils.getFromConfigPublicResponse
+import net.irisshaders.lilybot.utils.getConfigPublicResponse
 
 //todo This really just needs a full rework. I'm about 90% sure it's not properly adapted for cross guild work.
 class RoleMenu : Extension() {
@@ -70,7 +70,7 @@ class RoleMenu : Extension() {
 								arguments.role.id,
 								"add"
 							)
-							DatabaseHelper.putInComponents(newComponent)
+							DatabaseHelper.setComponent(newComponent)
 
 							action { }
 						}
@@ -85,7 +85,7 @@ class RoleMenu : Extension() {
 								arguments.role.id,
 								"remove"
 							)
-							DatabaseHelper.putInComponents(newComponent)
+							DatabaseHelper.setComponent(newComponent)
 
 							action { }
 						}
@@ -97,7 +97,7 @@ class RoleMenu : Extension() {
 
 					// Try to get the action log from the config.
 					// If a config is not set, inform the user and return@action
-					val actionLogId = getFromConfigPublicResponse("modActionLog") ?: return@action
+					val actionLogId = getConfigPublicResponse("modActionLog") ?: return@action
 					val actionLog = guild?.getChannel(actionLogId) as GuildMessageChannelBehavior
 
 					actionLog.createEmbed {
@@ -136,12 +136,12 @@ class RoleMenu : Extension() {
 
 				// this is  a very dirty fix, so it doesn't conflict with log uploading
 				val roleId: Snowflake? = try {
-					DatabaseHelper.selectInComponents(interaction.componentId, "roleId") as Snowflake?
+					DatabaseHelper.getComponent(interaction.componentId, "roleId") as Snowflake?
 				} catch (e: NullPointerException) {
 					return@action
 				}
 				val addOrRemove: String? = try {
-					DatabaseHelper.selectInComponents(interaction.componentId, "addOrRemove") as String?
+					DatabaseHelper.getComponent(interaction.componentId, "addOrRemove") as String?
 				} catch (e: NullPointerException) {
 					return@action
 				}
