+++ src/main/kotlin/net/irisshaders/lilybot/utils/_Utils.kt
@@ -14,8 +14,8 @@ import com.kotlindiscord.kord.extensions.types.respond
  * @return The column value from the database or null
  * @author NoComment1105
  */
-suspend fun EphemeralSlashCommandContext<*>.getFromConfigPrivateResponse(inputColumn: String) =
-	DatabaseHelper.selectInConfig(guild!!.id, inputColumn) ?: run {
+suspend fun EphemeralSlashCommandContext<*>.getConfigPrivateResponse(inputColumn: String) =
+	DatabaseHelper.getConfig(guild!!.id, inputColumn) ?: run {
 		respond {
 			content = "**Error:** Unable to access config for this guild! Is your configuration set?"
 		}
@@ -33,8 +33,8 @@ suspend fun EphemeralSlashCommandContext<*>.getFromConfigPrivateResponse(inputCo
  * @return The column value from the database or a public error and null
  * @author NoComment1105
  */
-suspend fun EphemeralSlashCommandContext<*>.getFromConfigPublicResponse(inputColumn: String) =
-	DatabaseHelper.selectInConfig(guild!!.id, inputColumn) ?: run {
+suspend fun EphemeralSlashCommandContext<*>.getConfigPublicResponse(inputColumn: String) =
+	DatabaseHelper.getConfig(guild!!.id, inputColumn) ?: run {
 		respond {
 			content = "**Error:** Unable to access config for this guild! Please inform a member of staff!"
 		}
@@ -51,8 +51,8 @@ suspend fun EphemeralSlashCommandContext<*>.getFromConfigPublicResponse(inputCol
  * @return The column value from the database or a private error and null
  * @author NoComment1105
  */
-suspend fun EphemeralMessageCommandContext.getFromConfigPublicResponse(inputColumn: String) =
-	DatabaseHelper.selectInConfig(guild!!.id, inputColumn) ?: run {
+suspend fun EphemeralMessageCommandContext.getConfigPublicResponse(inputColumn: String) =
+	DatabaseHelper.getConfig(guild!!.id, inputColumn) ?: run {
 		respond {
 			content = "**Error:** Unable to access config for this guild! Please inform a member of staff!"
 		}
